BACKLOG-3770 - 6.6 investigate purge versions tool for impvs https://jira.jahia.org/browse/BACKLOG-3770
Provide support in the NodeInfo to get the node type of the bundle (speeds up the lookup for nt:versionHitory nodes).
Added method to the BundleDbPersistenceManager to check if nodes are still references using a batch query.
Support deletion of the multiple versions and multiple version histories in one operation.

git-svn-id: https://devtools.jahia.com/svn/jackrabbit/trunk@117 25beda46-67d2-432a-828b-91af10607620