BACKLOG-3866 - Check if we can purge unused version using age or number of versions to keep https://jira.jahia.org/browse/BACKLOG-3866
Added support for removing unused versions (the check is done before the call). A new method, accepting a list of versions to remove is provided. This method does not perform the check for references (i.e. if the version is still used or not) and also does not perform the check if the removed version is the last one in the version history (except root) and the whole version history needs to be deleted. As we use this method for purging unused versions of existing nodes, the version history is still referenced. By bypassing reference checks the deletion is significantly faster. The check for references is done before in a build query.

git-svn-id: https://devtools.jahia.com/svn/jackrabbit/trunk@130 25beda46-67d2-432a-828b-91af10607620