JCR-3490 - More efficient node traversal during garbage collection https://issues.apache.org/jira/browse/JCR-3490
QA-3907 - OutOfMemoryError when DataStore Garbage Collector runs
https://jira.jahia.org/browse/QA-3907

Backport changes from Jackrabbit trunk:
https://fisheye6.atlassian.com/changelog/jackrabbit?cs=1351240
https://fisheye6.atlassian.com/changelog/jackrabbit?cs=1352849
https://fisheye6.atlassian.com/changelog/jackrabbit?cs=1427374

JCR-3263 load in structural node information immediately for better performance and add option to load the complete node graph first before checking for consistency. performance is boosted dozens of times on large repositories when enabling this option but requires more memory

JCR-3263: Consistency checker performance improvements
Add a default implementation of getAllNodeInfos() in AbstractBundlePersistenceManager to avoid having to reimplement the logic in all PMs.

JCR-3490 instead of individually loading every node in the repository load batches of node infos at once. node states are now only loaded when the node in question indeed has a blob property in the data store

Additionally: Do not use the NodeId cache in the NodeInfo as it is growing unpredictably and resulting in an OOM error.

git-svn-id: https://devtools.jahia.com/svn/jackrabbit/trunk@116 25beda46-67d2-432a-828b-91af10607620